[wip] move items to a separate module
