+++ src/lib.rs
@@ -10,1560 +10,15 @@
 #![cfg_attr(not(target_os = "windows"), doc=include_str!("../README.md"))]
 
 pub mod error;
+mod primitive_impls;
 mod private;
 pub mod traits;
+pub mod types;
 
 use error::Error;
-#[cfg(feature = "image")]
-use image::{DynamicImage, ImageFormat};
 pub use traits::{Read, ReadContext, ReadVersioned, Write};
+use types::{color::Colour, vec2::Vec2};
 use uuid::Uuid;
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
-struct Varint(i32);
-
-impl Read for Varint {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let value = leb128::read::signed(input)?;
-
-        Ok(Self(value.try_into().unwrap()))
-    }
-}
-
-impl Write for Varint {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        leb128::write::signed(output, self.0.into())?;
-
-        Ok(())
-    }
-}
-
-impl Read for String {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let len = Varint::read(input)?;
-
-        let mut string = Self::with_capacity(usize::try_from(len.0).unwrap());
-
-        for _ in 0..len.0 {
-            let c = u8::read(input)? as char;
-            string.push(c);
-        }
-
-        Ok(string)
-    }
-}
-
-impl Write for String {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        Varint(i32::try_from(self.len()).unwrap()).write(output)?;
-
-        for c in self.chars() {
-            (c as u8).write(output)?;
-        }
-
-        Ok(())
-    }
-}
-
-impl Write for u32 {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        Ok(output.write_all(&self.to_le_bytes())?)
-    }
-}
-
-impl Read for i32 {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let mut bytes = [0; 4];
-
-        for byte in &mut bytes {
-            *byte = Read::read(input)?;
-        }
-
-        Ok(Self::from_le_bytes(bytes))
-    }
-}
-
-impl Write for i32 {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        Ok(output.write_all(&self.to_le_bytes())?)
-    }
-}
-
-impl Read for i64 {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let mut bytes = [0; 8];
-
-        for byte in &mut bytes {
-            *byte = Read::read(input)?;
-        }
-
-        Ok(Self::from_le_bytes(bytes))
-    }
-}
-
-impl Write for i64 {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        Ok(output.write_all(&self.to_le_bytes())?)
-    }
-}
-
-impl Read for f32 {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let mut bytes = [0; 4];
-
-        for byte in &mut bytes {
-            *byte = Read::read(input)?;
-        }
-
-        Ok(Self::from_le_bytes(bytes))
-    }
-}
-
-impl Write for f32 {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        Ok(output.write_all(&self.to_le_bytes())?)
-    }
-}
-
-impl<T: Read> Read for Vec<T> {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let len = usize::try_from(i32::read(input)?).unwrap();
-
-        let mut vec = Self::with_capacity(len);
-
-        for _ in 0..len {
-            vec.push(Read::read(input)?);
-        }
-
-        Ok(vec)
-    }
-}
-
-impl<T: Write> Write for Vec<T> {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        i32::try_from(self.len()).unwrap().write(output)?;
-
-        for item in self {
-            item.write(output)?;
-        }
-
-        Ok(())
-    }
-}
-
-impl<T: Read + Copy + Default, const LEN: usize> Read for [T; LEN] {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let mut arr = [Default::default(); LEN];
-
-        for item in &mut arr {
-            *item = Read::read(input)?;
-        }
-
-        Ok(arr)
-    }
-}
-
-impl<T: Write, const LEN: usize> Write for [T; LEN] {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        for item in self {
-            item.write(output)?;
-        }
-
-        Ok(())
-    }
-}
-
-impl<T: Read> Read for Option<T> {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        if bool::read(input)? {
-            Ok(Some(Read::read(input)?))
-        } else {
-            Ok(None)
-        }
-    }
-}
-
-impl<T: Write> Write for Option<T> {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.is_some().write(output)?;
-
-        if let Some(value) = self {
-            value.write(output)?;
-        }
-
-        Ok(())
-    }
-}
-
-impl Read for bool {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(u8::read(input)? != 0)
-    }
-}
-
-impl Write for bool {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        u8::from(*self).write(output)
-    }
-}
-
-impl Read for u8 {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let mut buf = [0; 1];
-        input.read_exact(&mut buf)?;
-        Ok(buf[0])
-    }
-}
-
-impl Write for u8 {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        Ok(output.write_all(&[*self])?)
-    }
-}
-
-#[cfg(feature = "image")]
-impl Read for DynamicImage {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let vec = Vec::<u8>::read(input)?;
-
-        image::load_from_memory(&vec).map_err(Error::from)
-    }
-}
-
-#[cfg(feature = "image")]
-impl Write for DynamicImage {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        let mut vec = std::io::Cursor::new(Vec::new());
-        self.write_to(&mut vec, ImageFormat::Png)?;
-
-        output.write_all(&vec.into_inner())?;
-
-        Ok(())
-    }
-}
-
-/// A full Exoracer level.
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug)]
-pub struct Exolvl {
-    /// The local level data for this level.
-    pub local_level: LocalLevel,
-    /// The actual level data.
-    pub level_data: LevelData,
-    /// The data for the author time replay.
-    pub author_replay: AuthorReplay,
-}
-
-const EXPECTED_MAGIC: &[u8; 4] = b"NYA^";
-
-impl Read for Exolvl {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let magic: [u8; 4] = Read::read(input)?;
-
-        if &magic != EXPECTED_MAGIC {
-            return Err(Error::WrongMagic);
-        }
-
-        let local_level = LocalLevel::read(input)?;
-        let level_data = ReadVersioned::read(input, local_level.serialization_version)?;
-        let author_replay = Read::read(input)?;
-
-        Ok(Self {
-            local_level,
-            level_data,
-            author_replay,
-        })
-    }
-}
-
-impl Write for Exolvl {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        EXPECTED_MAGIC.write(output)?;
-        self.local_level.write(output)?;
-        self.level_data.write(output)?;
-        self.author_replay.write(output)
-    }
-}
-
-impl Default for Exolvl {
-    fn default() -> Self {
-        let level_id = uuid::Uuid::new_v4();
-        Self {
-            local_level: LocalLevel {
-                serialization_version: 18,
-                level_id,
-                level_version: 1,
-                level_name: "New level".to_string(),
-                thumbnail: String::default(),
-                creation_date: chrono::Utc::now(),
-                update_date: chrono::Utc::now(),
-                author_time: Default::default(),
-                author_lap_times: Vec::default(),
-                silver_medal_time: Default::default(),
-                gold_medal_time: Default::default(),
-                laps: 1,
-                private: Default::default(),
-                nova_level: true,
-            },
-            level_data: LevelData {
-                level_id,
-                level_version: 1,
-                nova_level: true,
-                under_decoration_tiles: Vec::default(),
-                background_decoration_tiles: Vec::default(),
-                terrain_tiles: Vec::default(),
-                floating_zone_tiles: Vec::default(),
-                object_tiles: Vec::default(),
-                foreground_decoration_tiles: Vec::default(),
-                objects: Vec::default(),
-                layers: Vec::default(),
-                prefabs: Vec::default(),
-                brushes: Vec::default(),
-                patterns: Vec::default(),
-                colour_palette: Some(Vec::default()),
-                author_time: Default::default(),
-                author_lap_times: Vec::default(),
-                silver_medal_time: Default::default(),
-                gold_medal_time: Default::default(),
-                laps: 1,
-                center_camera: Default::default(),
-                scripts: Vec::default(),
-                nova_scripts: Vec::default(),
-                global_variables: Vec::default(),
-                theme: "mountains".to_string(),
-                custom_background_colour: Colour::default(),
-                unknown1: [0; 4],
-                custom_terrain_pattern_id: Default::default(),
-                custom_terrain_pattern_tiling: Vec2::default(),
-                custom_terrain_pattern_offset: Vec2::default(),
-                custom_terrain_colour: Colour::default(),
-                custom_terrain_secondary_color: Colour::default(),
-                custom_terrain_blend_mode: Default::default(),
-                custom_terrain_border_colour: Colour::default(),
-                custom_terrain_border_thickness: Default::default(),
-                custom_terrain_border_corner_radius: Default::default(),
-                custom_terrain_round_reflex_angles: Default::default(),
-                custom_terrain_round_collider: Default::default(),
-                custom_terrain_friction: Default::default(),
-                default_music: true,
-                music_ids: Vec::default(),
-                allow_direction_change: Default::default(),
-                disable_replays: Default::default(),
-                disable_revive_pads: Default::default(),
-                disable_start_animation: Default::default(),
-                gravity: Vec2 { x: 0.0, y: -75.0 },
-            },
-            author_replay: AuthorReplay(Vec::default()),
-        }
-    }
-}
-/// The local level data for this level.
-///
-/// This data is only ever used in the level editor and is not uploaded to the server.
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug, Hash)]
-pub struct LocalLevel {
-    /// The version of the exolvl format that this level uses.
-    ///
-    /// The current latest serialization version is 18.
-    pub serialization_version: i32,
-    /// The UUID of the level.
-    pub level_id: Uuid,
-    /// The version of the level e.g. v1, v2, etc.
-    pub level_version: i32,
-    /// The name of the level.
-    pub level_name: String,
-    /// The base64-encoded data for the thumbnail of the level.
-    pub thumbnail: String,
-    /// When this level was created.
-    pub creation_date: chrono::DateTime<chrono::Utc>,
-    /// When this level was last updated.
-    pub update_date: chrono::DateTime<chrono::Utc>,
-    /// The author medal time for this level in milliseconds.
-    pub author_time: i64,
-    /// The author medal lap times for this level in milliseconds.
-    pub author_lap_times: Vec<i64>,
-    /// The silver medal time for this level in milliseconds.
-    pub silver_medal_time: i64,
-    /// The gold medal time for this level in milliseconds.
-    pub gold_medal_time: i64,
-    /// The number of laps in this level.
-    pub laps: i32,
-    /// Whether this level is private or public.
-    pub private: bool,
-
-    /// If this is true, the level can be opened in the new level editor. Otherwise it's for the "legacy" editor.
-    pub nova_level: bool,
-}
-
-impl Read for LocalLevel {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self {
-            serialization_version: Read::read(input)?,
-            level_id: Read::read(input)?,
-            level_version: Read::read(input)?,
-            level_name: Read::read(input)?,
-            thumbnail: Read::read(input)?,
-            creation_date: Read::read(input)?,
-            update_date: Read::read(input)?,
-            author_time: Read::read(input)?,
-            author_lap_times: Read::read(input)?,
-            silver_medal_time: Read::read(input)?,
-            gold_medal_time: Read::read(input)?,
-            laps: Read::read(input)?,
-            private: Read::read(input)?,
-            nova_level: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for LocalLevel {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.serialization_version.write(output)?;
-        self.level_id.write(output)?;
-        self.level_version.write(output)?;
-        self.level_name.write(output)?;
-        self.thumbnail.write(output)?;
-        self.creation_date.write(output)?;
-        self.update_date.write(output)?;
-        self.author_time.write(output)?;
-        self.author_lap_times.write(output)?;
-        self.silver_medal_time.write(output)?;
-        self.gold_medal_time.write(output)?;
-        self.laps.write(output)?;
-        self.private.write(output)?;
-        self.nova_level.write(output)
-    }
-}
-
-const TICKS_TO_SECONDS: i64 = 10_000_000;
-const EPOCH_DIFFERENCE: i64 = 62_135_596_800;
-
-impl Read for chrono::DateTime<chrono::Utc> {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let ticks = i64::read(input)?;
-
-        let seconds = ticks / TICKS_TO_SECONDS - EPOCH_DIFFERENCE;
-
-        Ok(Self::from_timestamp(seconds, 0).unwrap())
-    }
-}
-
-impl Write for chrono::DateTime<chrono::Utc> {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        let ticks = (self.timestamp() + EPOCH_DIFFERENCE) * TICKS_TO_SECONDS;
-
-        ticks.write(output)
-    }
-}
-
-/// The level data for an Exoracer level.
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug)]
-pub struct LevelData {
-    /// The UUID of the level.
-    pub level_id: Uuid,
-    /// The version of the level e.g. v1, v2, etc.
-    pub level_version: i32,
-    /// Whether this level is for the new level editor.
-    ///
-    /// If this is true, the level can be opened in the new level editor. Otherwise it's for the "legacy" editor.
-    /// This Field is presumably only useful in .level files, not in .exolvl ones. A mismatch with the corresponding `LocalLevel` field should be avoided.
-    pub nova_level: bool,
-    /// The tile ids for the "under decoration" layer.
-    pub under_decoration_tiles: Vec<i32>,
-    /// The tile ids for the "background decoration" layer.
-    pub background_decoration_tiles: Vec<i32>,
-    /// The tile ids for the terrain layer.
-    pub terrain_tiles: Vec<i32>,
-    /// The tile ids for the floating zone layer.
-    pub floating_zone_tiles: Vec<i32>,
-    /// The tile ids for the "object" layer.
-    pub object_tiles: Vec<i32>,
-    /// The tile ids for the "foreground decoration" layer.
-    pub foreground_decoration_tiles: Vec<i32>,
-    /// The objects in the level.
-    pub objects: Vec<Object>,
-    /// The layers in the level.
-    pub layers: Vec<Layer>,
-    /// The prefabs in the level.
-    pub prefabs: Vec<Prefab>,
-    /// The brushes in the level.
-    pub brushes: Vec<Brush>,
-    /// The patterns in the level.
-    pub patterns: Vec<Pattern>,
-    /// The colour palettes in the level.
-    ///
-    /// This is only present in levels with version 17 or higher.
-    pub colour_palette: Option<Vec<Colour>>,
-    /// The author medal time for this level in milliseconds.
-    pub author_time: i64,
-    /// The author medal lap times for this level in milliseconds.
-    pub author_lap_times: Vec<i64>,
-    /// The silver medal time for this level in milliseconds.
-    pub silver_medal_time: i64,
-    /// The gold medal time for this level in milliseconds.
-    pub gold_medal_time: i64,
-    /// The number of laps in this level.
-    pub laps: i32,
-    /// Whether the camera should be centered while playing this level.
-    ///
-    /// This is mostly deprecated and should stay set to false.
-    pub center_camera: bool,
-    /// The scripts in the level.
-    ///
-    /// These are used in the legacy level editor.
-    pub scripts: Vec<i32>,
-    /// The "new" scripts in the level.
-    ///
-    /// These are the scripts that are used in the new level editor. As opposed to the `scripts` field which is for the legacy editor.
-    pub nova_scripts: Vec<NovaScript>,
-    /// All the global variables in the level.
-    pub global_variables: Vec<Variable>,
-    /// The theme name of the level.
-    pub theme: String,
-    /// The custom background colour of the level.
-    pub custom_background_colour: Colour,
-
-    /// Unknown data.
-    unknown1: [u8; 4],
-    /// The following terrain related fields are all used when explicitly copying certain terrain data.
-    ///
-    /// The custom terrain pattern that can be pasted with the `colour_paste` button if the recieving object has the `FillMode` set to `Pattern`.
-    pub custom_terrain_pattern_id: i32,
-    /// The tiling of that pattern.
-    pub custom_terrain_pattern_tiling: Vec2,
-    /// the offset of that pattern.
-    pub custom_terrain_pattern_offset: Vec2,
-    /// In the legacy editor: The custom terrain colour of the level.
-    /// In the new editor: The colour of the copied terrain.
-    pub custom_terrain_colour: Colour,
-    /// Not 100% sure of the use of this, presumably the replacement for the border color in the new editor.
-    /// Used when copying and pasting properties of terrain.
-    pub custom_terrain_secondary_color: Colour,
-    /// The blend mode of the copied terrain.
-    pub custom_terrain_blend_mode: i32,
-    /// The custom terrain border colour of the level.
-    pub custom_terrain_border_colour: Colour,
-    /// The thickness of the terrain border.
-    pub custom_terrain_border_thickness: f32,
-    /// The corner radius of the terrain border.
-    pub custom_terrain_border_corner_radius: f32,
-    /// Whether the copied terrain has round reflex angles or not (only visual).
-    pub custom_terrain_round_reflex_angles: bool,
-    /// Whether the copied terrain has a round collider or not (not visual).
-    pub custom_terrain_round_collider: bool,
-    /// The friction of the copied terrain.
-    pub custom_terrain_friction: f32,
-    /// Whether the default music should be played or not.
-    pub default_music: bool,
-    /// The music ids for the level. The game randomly picks one of these to play each time.
-    pub music_ids: Vec<String>,
-    /// Whether the game lets the player change directions or not.
-    pub allow_direction_change: bool,
-    /// Whether replays are disabled or not.
-    ///
-    /// If this is true, the game won't upload replays on this level.
-    /// (unless you explicitly upload blank shells from the history section, that only contain the time you set without any replay data. Could be a bug).
-    pub disable_replays: bool,
-    /// Whether revive pads are disabled or not.
-    pub disable_revive_pads: bool,
-    /// Whether the start animation is disabled or not.
-    pub disable_start_animation: bool,
-    /// The gravity vector for this level.
-    pub gravity: Vec2,
-}
-
-impl ReadVersioned for LevelData {
-    fn read(input: &mut impl std::io::Read, version: i32) -> Result<Self, Error> {
-        Ok(Self {
-            level_id: Read::read(input)?,
-            level_version: Read::read(input)?,
-            nova_level: Read::read(input)?,
-            under_decoration_tiles: Read::read(input)?,
-            background_decoration_tiles: Read::read(input)?,
-            terrain_tiles: Read::read(input)?,
-            floating_zone_tiles: Read::read(input)?,
-            object_tiles: Read::read(input)?,
-            foreground_decoration_tiles: Read::read(input)?,
-            objects: Read::read(input)?,
-            layers: Read::read(input)?,
-            prefabs: Read::read(input)?,
-            brushes: Read::read(input)?,
-            patterns: Read::read(input)?,
-            colour_palette: if version >= 17 {
-                Some(Read::read(input)?)
-            } else {
-                None
-            },
-            author_time: Read::read(input)?,
-            author_lap_times: Read::read(input)?,
-            silver_medal_time: Read::read(input)?,
-            gold_medal_time: Read::read(input)?,
-            laps: Read::read(input)?,
-            center_camera: Read::read(input)?,
-            scripts: Read::read(input)?,
-            nova_scripts: Read::read(input)?,
-            global_variables: Read::read(input)?,
-            theme: Read::read(input)?,
-            custom_background_colour: Read::read(input)?,
-            unknown1: Read::read(input)?,
-            custom_terrain_pattern_id: Read::read(input)?,
-            custom_terrain_pattern_tiling: Read::read(input)?,
-            custom_terrain_pattern_offset: Read::read(input)?,
-            custom_terrain_colour: Read::read(input)?,
-            custom_terrain_secondary_color: Read::read(input)?,
-            custom_terrain_blend_mode: Read::read(input)?,
-            custom_terrain_border_colour: Read::read(input)?,
-            custom_terrain_border_thickness: Read::read(input)?,
-            custom_terrain_border_corner_radius: Read::read(input)?,
-            custom_terrain_round_reflex_angles: Read::read(input)?,
-            custom_terrain_round_collider: Read::read(input)?,
-            custom_terrain_friction: Read::read(input)?,
-            default_music: Read::read(input)?,
-            music_ids: Read::read(input)?,
-            allow_direction_change: Read::read(input)?,
-            disable_replays: Read::read(input)?,
-            disable_revive_pads: Read::read(input)?,
-            disable_start_animation: Read::read(input)?,
-            gravity: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for LevelData {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.level_id.write(output)?;
-        self.level_version.write(output)?;
-        self.nova_level.write(output)?;
-        self.under_decoration_tiles.write(output)?;
-        self.background_decoration_tiles.write(output)?;
-        self.terrain_tiles.write(output)?;
-        self.floating_zone_tiles.write(output)?;
-        self.object_tiles.write(output)?;
-        self.foreground_decoration_tiles.write(output)?;
-        self.objects.write(output)?;
-        self.layers.write(output)?;
-        self.prefabs.write(output)?;
-        self.brushes.write(output)?;
-        self.patterns.write(output)?;
-        if let Some(colour_palette) = &self.colour_palette {
-            colour_palette.write(output)?;
-        }
-        self.author_time.write(output)?;
-        self.author_lap_times.write(output)?;
-        self.silver_medal_time.write(output)?;
-        self.gold_medal_time.write(output)?;
-        self.laps.write(output)?;
-        self.center_camera.write(output)?;
-        self.scripts.write(output)?;
-        self.nova_scripts.write(output)?;
-        self.global_variables.write(output)?;
-        self.theme.write(output)?;
-        self.custom_background_colour.write(output)?;
-        self.unknown1.write(output)?;
-        self.custom_terrain_pattern_id.write(output)?;
-        self.custom_terrain_pattern_tiling.write(output)?;
-        self.custom_terrain_pattern_offset.write(output)?;
-        self.custom_terrain_colour.write(output)?;
-        self.custom_terrain_secondary_color.write(output)?;
-        self.custom_terrain_blend_mode.write(output)?;
-        self.custom_terrain_border_colour.write(output)?;
-        self.custom_terrain_border_thickness.write(output)?;
-        self.custom_terrain_border_corner_radius.write(output)?;
-        self.custom_terrain_round_reflex_angles.write(output)?;
-        self.custom_terrain_round_collider.write(output)?;
-        self.custom_terrain_friction.write(output)?;
-        self.default_music.write(output)?;
-        self.music_ids.write(output)?;
-        self.allow_direction_change.write(output)?;
-        self.disable_replays.write(output)?;
-        self.disable_revive_pads.write(output)?;
-        self.disable_start_animation.write(output)?;
-        self.gravity.write(output)
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "image", derive(Clone, Debug, PartialEq))]
-#[cfg_attr(
-    not(feature = "image"),
-    derive(Clone, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)
-)]
-pub struct Pattern {
-    pub pattern_id: i32,
-    pub pattern_frames: Vec<Image>,
-}
-
-impl Read for Pattern {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self {
-            pattern_id: Read::read(input)?,
-            pattern_frames: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for Pattern {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.pattern_id.write(output)?;
-        self.pattern_frames.write(output)
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug)]
-pub struct Prefab {
-    pub prefab_id: i32,
-    pub prefab_image_data: Image,
-    pub items: Vec<Object>,
-}
-
-impl Read for Prefab {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self {
-            prefab_id: Read::read(input)?,
-            prefab_image_data: Read::read(input)?,
-            items: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for Prefab {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.prefab_id.write(output)?;
-        self.prefab_image_data.write(output)?;
-        self.items.write(output)
-    }
-}
-
-#[cfg(feature = "image")]
-#[derive(Clone, Debug, PartialEq)]
-pub struct Image(pub DynamicImage);
-
-#[cfg(all(feature = "image", feature = "serde"))]
-impl serde::Serialize for Image {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        self.0.as_bytes().serialize(serializer)
-    }
-}
-
-#[cfg(all(feature = "image", feature = "serde"))]
-impl<'de> serde::Deserialize<'de> for Image {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let buffer = serde::Deserialize::deserialize(deserializer)?;
-
-        let img = image::load_from_memory(buffer).map_err(serde::de::Error::custom)?;
-
-        Ok(Self(img))
-    }
-}
-
-#[cfg(not(feature = "image"))]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub struct Image(pub Vec<u8>);
-
-impl Read for Image {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let data = Read::read(input)?;
-
-        Ok(Self(data))
-    }
-}
-
-impl Write for Image {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.0.write(output)
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug)]
-pub struct Layer {
-    pub layer_id: i32,
-    pub layer_name: String,
-    pub selected: bool,
-    pub invisible: bool,
-    pub locked: bool,
-    pub foreground_type: i32,
-    pub parallax: Vec2,
-    pub fixed_size: bool,
-    pub children: Vec<i32>,
-}
-
-impl Read for Layer {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self {
-            layer_id: Read::read(input)?,
-            layer_name: Read::read(input)?,
-            selected: Read::read(input)?,
-            invisible: Read::read(input)?,
-            locked: Read::read(input)?,
-            foreground_type: Read::read(input)?,
-            parallax: Read::read(input)?,
-            fixed_size: Read::read(input)?,
-            children: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for Layer {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.layer_id.write(output)?;
-        self.layer_name.write(output)?;
-        self.selected.write(output)?;
-        self.invisible.write(output)?;
-        self.locked.write(output)?;
-        self.foreground_type.write(output)?;
-        self.parallax.write(output)?;
-        self.fixed_size.write(output)?;
-        self.children.write(output)
-    }
-}
-
-/// A 2D vector.
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Copy, Debug)]
-pub struct Vec2 {
-    /// The x-coordinate.
-    pub x: f32,
-    /// The y-coordinate.
-    pub y: f32,
-}
-
-impl Read for Vec2 {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self {
-            x: Read::read(input)?,
-            y: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for Vec2 {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.x.write(output)?;
-        self.y.write(output)
-    }
-}
-
-impl Default for Vec2 {
-    fn default() -> Self {
-        Self { x: 0.0, y: 0.0 }
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Copy, Debug)]
-pub struct Colour {
-    pub r: f32,
-    pub g: f32,
-    pub b: f32,
-    pub a: f32,
-}
-
-impl Read for Colour {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self {
-            r: Read::read(input)?,
-            g: Read::read(input)?,
-            b: Read::read(input)?,
-            a: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for Colour {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.r.write(output)?;
-        self.g.write(output)?;
-        self.b.write(output)?;
-        self.a.write(output)
-    }
-}
-
-impl Default for Colour {
-    fn default() -> Self {
-        Self {
-            r: 0.0,
-            g: 0.0,
-            b: 0.0,
-            a: 1.0,
-        }
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub struct AuthorReplay(pub Vec<u8>);
-
-impl Read for AuthorReplay {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self(Read::read(input)?))
-    }
-}
-
-impl Write for AuthorReplay {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.0.write(output)
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug)]
-pub struct Object {
-    pub entity_id: i32,
-    pub tile_id: i32,
-    pub prefab_entity_id: i32,
-    pub prefab_id: i32,
-    pub position: Vec2,
-    pub scale: Vec2,
-    pub rotation: f32,
-    pub tag: String,
-    pub properties: Vec<ObjectProperty>,
-    pub in_layer: i32,
-    pub in_group: i32,
-    pub group_members: Vec<i32>,
-}
-
-impl Read for Object {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self {
-            entity_id: Read::read(input)?,
-            tile_id: Read::read(input)?,
-            prefab_entity_id: Read::read(input)?,
-            prefab_id: Read::read(input)?,
-            position: Read::read(input)?,
-            scale: Read::read(input)?,
-            rotation: Read::read(input)?,
-            tag: Read::read(input)?,
-            properties: Read::read(input)?,
-            in_layer: Read::read(input)?,
-            in_group: Read::read(input)?,
-            group_members: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for Object {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.entity_id.write(output)?;
-        self.tile_id.write(output)?;
-        self.prefab_entity_id.write(output)?;
-        self.prefab_id.write(output)?;
-        self.position.write(output)?;
-        self.scale.write(output)?;
-        self.rotation.write(output)?;
-        self.tag.write(output)?;
-        self.properties.write(output)?;
-        self.in_layer.write(output)?;
-        self.in_group.write(output)?;
-        self.group_members.write(output)
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug)]
-pub enum ObjectProperty {
-    Colour(Colour),
-    Resolution(i32),
-    FillMode(i32),
-    SecondaryColour(Colour),
-    Thickness(f32),
-    TotalAngle(i32),
-    Corners(i32),
-    Blending(i32),
-    GridOffset(Vec2),
-    CornerRadius(f32),
-    Width(f32),
-    Height(f32),
-    BorderColour(Colour),
-    BorderThickness(f32),
-    PhysicsType(i32),
-    Friction(f32),
-    TerrainCorners(Vec<Vec<Vec2>>),
-    Direction(i32),
-    Impulse(i32),
-    Killer(bool),
-    RoundReflexAngles(bool),
-    RoundCollider(bool),
-    Radius(f32),
-    Size(f32),
-    ReverseDirection(bool),
-    CollisionDetector(bool),
-    Pattern(i32),
-    PatternTiling(Vec2),
-    PatternOffset(Vec2),
-    Bounce(bool),
-    RestoreVelocity(bool),
-    Sprite(String),
-    Trigger(bool),
-    Health(f32),
-    DamageFromJump(bool),
-    DamageFromDash(bool),
-    ReverseDirOnDamage(bool),
-    Floating(bool),
-    LinkedObjects(Vec<i32>),
-    FlipX(bool),
-    FlipY(bool),
-    Text(String),
-    FontSize(f32),
-    EditorColour(Colour),
-    Colour2(Colour),
-    Colour3(Colour),
-    Colour4(Colour),
-    ParticleTexture(String),
-    Duration(f32),
-    Delay(f32),
-    Loop(bool),
-    AutoPlay(bool),
-    LifetimeMin(f32),
-    LifetimeMax(f32),
-    SimulationSpace(i32),
-    Rate(f32),
-    Burst(i32),
-    EmitterShape(i32),
-    EmitterWidth(f32),
-    EmitterHeight(f32),
-    EmitterTotalAngle(f32),
-    SizeMin(f32),
-    SizeMax(f32),
-    SizeOverLifetime(bool),
-    StartSizeMultiplier(f32),
-    EndSizeMultiplier(f32),
-    SpeedMin(f32),
-    SpeedMax(f32),
-    SpeeLimit(f32),
-    SpeedDampen(f32),
-    RotationMin(f32),
-    RotationMax(f32),
-    Rotationspeed(f32),
-    ColourOverLifetime(bool),
-    StartColourMultiplier(Colour),
-    EndColourMultiplier(Colour),
-    GravityMultiplier(f32),
-    AnchorPos(Vec2),
-    MoonInnerRadius(f32),
-    MoonOffset(f32),
-}
-
-impl Read for ObjectProperty {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        let property_type = Read::read(input)?;
-
-        Ok(match property_type {
-            0 => Self::Colour(Read::read(input)?),
-            1 => Self::Resolution(Read::read(input)?),
-            2 => Self::FillMode(Read::read(input)?),
-            3 => Self::SecondaryColour(Read::read(input)?),
-            4 => Self::Thickness(Read::read(input)?),
-            5 => Self::TotalAngle(Read::read(input)?),
-            6 => Self::Corners(Read::read(input)?),
-            7 => Self::Blending(Read::read(input)?),
-            8 => Self::GridOffset(Read::read(input)?),
-            9 => Self::CornerRadius(Read::read(input)?),
-            10 => Self::Width(Read::read(input)?),
-            11 => Self::Height(Read::read(input)?),
-            12 => Self::BorderColour(Read::read(input)?),
-            13 => Self::BorderThickness(Read::read(input)?),
-            14 => Self::PhysicsType(Read::read(input)?),
-            15 => Self::Friction(Read::read(input)?),
-            16 => Self::TerrainCorners(Read::read(input)?),
-            17 => Self::Direction(Read::read(input)?),
-            18 => Self::Impulse(Read::read(input)?),
-            19 => Self::Killer(Read::read(input)?),
-            20 => Self::RoundReflexAngles(Read::read(input)?),
-            21 => Self::RoundCollider(Read::read(input)?),
-            22 => Self::Radius(Read::read(input)?),
-            23 => Self::Size(Read::read(input)?),
-            24 => Self::ReverseDirection(Read::read(input)?),
-            25 => Self::CollisionDetector(Read::read(input)?),
-            26 => Self::Pattern(Read::read(input)?),
-            27 => Self::PatternTiling(Read::read(input)?),
-            28 => Self::PatternOffset(Read::read(input)?),
-            32 => Self::Bounce(Read::read(input)?),
-            34 => Self::RestoreVelocity(Read::read(input)?),
-            35 => Self::Sprite(Read::read(input)?),
-            36 => Self::Trigger(Read::read(input)?),
-            37 => Self::Health(Read::read(input)?),
-            38 => Self::DamageFromJump(Read::read(input)?),
-            39 => Self::DamageFromDash(Read::read(input)?),
-            40 => Self::ReverseDirOnDamage(Read::read(input)?),
-            41 => Self::Floating(Read::read(input)?),
-            42 => Self::LinkedObjects(Read::read(input)?),
-            43 => Self::FlipX(Read::read(input)?),
-            44 => Self::FlipY(Read::read(input)?),
-            45 => Self::Text(Read::read(input)?),
-            46 => Self::FontSize(Read::read(input)?),
-            47 => Self::EditorColour(Read::read(input)?),
-            48 => Self::Colour2(Read::read(input)?),
-            49 => Self::Colour3(Read::read(input)?),
-            50 => Self::Colour4(Read::read(input)?),
-            51 => Self::ParticleTexture(Read::read(input)?),
-            52 => Self::Duration(Read::read(input)?),
-            53 => Self::Delay(Read::read(input)?),
-            54 => Self::Loop(Read::read(input)?),
-            55 => Self::AutoPlay(Read::read(input)?),
-            56 => Self::LifetimeMin(Read::read(input)?),
-            57 => Self::LifetimeMax(Read::read(input)?),
-            58 => Self::SimulationSpace(Read::read(input)?),
-            59 => Self::Rate(Read::read(input)?),
-            60 => Self::Burst(Read::read(input)?),
-            61 => Self::EmitterShape(Read::read(input)?),
-            62 => Self::EmitterWidth(Read::read(input)?),
-            63 => Self::EmitterHeight(Read::read(input)?),
-            64 => Self::EmitterTotalAngle(Read::read(input)?),
-            65 => Self::SizeMin(Read::read(input)?),
-            66 => Self::SizeMax(Read::read(input)?),
-            67 => Self::SizeOverLifetime(Read::read(input)?),
-            68 => Self::StartSizeMultiplier(Read::read(input)?),
-            69 => Self::EndSizeMultiplier(Read::read(input)?),
-            71 => Self::SpeedMin(Read::read(input)?),
-            72 => Self::SpeedMax(Read::read(input)?),
-            73 => Self::SpeeLimit(Read::read(input)?),
-            74 => Self::SpeedDampen(Read::read(input)?),
-            75 => Self::RotationMin(Read::read(input)?),
-            76 => Self::RotationMax(Read::read(input)?),
-            77 => Self::Rotationspeed(Read::read(input)?),
-            78 => Self::ColourOverLifetime(Read::read(input)?),
-            79 => Self::StartColourMultiplier(Read::read(input)?),
-            80 => Self::EndColourMultiplier(Read::read(input)?),
-            81 => Self::GravityMultiplier(Read::read(input)?),
-            82 => Self::AnchorPos(Read::read(input)?),
-            83 => Self::MoonInnerRadius(Read::read(input)?),
-            84 => Self::MoonOffset(Read::read(input)?),
-            n => return Err(crate::error::Error::InvalidObjectPropertyType(n)),
-        })
-    }
-}
-
-impl Write for ObjectProperty {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        match self {
-            Self::Colour(value) => {
-                0.write(output)?;
-                value.write(output)
-            }
-            Self::Resolution(value) => {
-                1.write(output)?;
-                value.write(output)
-            }
-            Self::FillMode(value) => {
-                2.write(output)?;
-                value.write(output)
-            }
-            Self::SecondaryColour(value) => {
-                3.write(output)?;
-                value.write(output)
-            }
-            Self::Thickness(value) => {
-                4.write(output)?;
-                value.write(output)
-            }
-            Self::TotalAngle(value) => {
-                5.write(output)?;
-                value.write(output)
-            }
-            Self::Corners(value) => {
-                6.write(output)?;
-                value.write(output)
-            }
-            Self::Blending(value) => {
-                7.write(output)?;
-                value.write(output)
-            }
-            Self::GridOffset(value) => {
-                8.write(output)?;
-                value.write(output)
-            }
-            Self::CornerRadius(value) => {
-                9.write(output)?;
-                value.write(output)
-            }
-            Self::Width(value) => {
-                10.write(output)?;
-                value.write(output)
-            }
-            Self::Height(value) => {
-                11.write(output)?;
-                value.write(output)
-            }
-            Self::BorderColour(value) => {
-                12.write(output)?;
-                value.write(output)
-            }
-            Self::BorderThickness(value) => {
-                13.write(output)?;
-                value.write(output)
-            }
-            Self::PhysicsType(value) => {
-                14.write(output)?;
-                value.write(output)
-            }
-            Self::Friction(value) => {
-                15.write(output)?;
-                value.write(output)
-            }
-            Self::TerrainCorners(value) => {
-                16.write(output)?;
-                value.write(output)
-            }
-            Self::Direction(value) => {
-                17.write(output)?;
-                value.write(output)
-            }
-            Self::Impulse(value) => {
-                18.write(output)?;
-                value.write(output)
-            }
-            Self::Killer(value) => {
-                19.write(output)?;
-                value.write(output)
-            }
-            Self::RoundReflexAngles(value) => {
-                20.write(output)?;
-                value.write(output)
-            }
-            Self::RoundCollider(value) => {
-                21.write(output)?;
-                value.write(output)
-            }
-            Self::Radius(value) => {
-                22.write(output)?;
-                value.write(output)
-            }
-            Self::Size(value) => {
-                23.write(output)?;
-                value.write(output)
-            }
-            Self::ReverseDirection(value) => {
-                24.write(output)?;
-                value.write(output)
-            }
-            Self::CollisionDetector(value) => {
-                25.write(output)?;
-                value.write(output)
-            }
-            Self::Pattern(value) => {
-                26.write(output)?;
-                value.write(output)
-            }
-            Self::PatternTiling(value) => {
-                27.write(output)?;
-                value.write(output)
-            }
-            Self::PatternOffset(value) => {
-                28.write(output)?;
-                value.write(output)
-            }
-            Self::Bounce(value) => {
-                32.write(output)?;
-                value.write(output)
-            }
-            Self::RestoreVelocity(value) => {
-                34.write(output)?;
-                value.write(output)
-            }
-            Self::Sprite(value) => {
-                35.write(output)?;
-                value.write(output)
-            }
-            Self::Trigger(value) => {
-                36.write(output)?;
-                value.write(output)
-            }
-            Self::Health(value) => {
-                37.write(output)?;
-                value.write(output)
-            }
-            Self::DamageFromJump(value) => {
-                38.write(output)?;
-                value.write(output)
-            }
-            Self::DamageFromDash(value) => {
-                39.write(output)?;
-                value.write(output)
-            }
-            Self::ReverseDirOnDamage(value) => {
-                40.write(output)?;
-                value.write(output)
-            }
-            Self::Floating(value) => {
-                41.write(output)?;
-                value.write(output)
-            }
-            Self::LinkedObjects(value) => {
-                42.write(output)?;
-                value.write(output)
-            }
-            Self::FlipX(value) => {
-                43.write(output)?;
-                value.write(output)
-            }
-            Self::FlipY(value) => {
-                44.write(output)?;
-                value.write(output)
-            }
-            Self::Text(value) => {
-                45.write(output)?;
-                value.write(output)
-            }
-            Self::FontSize(value) => {
-                46.write(output)?;
-                value.write(output)
-            }
-            Self::EditorColour(value) => {
-                47.write(output)?;
-                value.write(output)
-            }
-            Self::Colour2(value) => {
-                48.write(output)?;
-                value.write(output)
-            }
-            Self::Colour3(value) => {
-                49.write(output)?;
-                value.write(output)
-            }
-            Self::Colour4(value) => {
-                50.write(output)?;
-                value.write(output)
-            }
-            Self::ParticleTexture(value) => {
-                51.write(output)?;
-                value.write(output)
-            }
-            Self::Duration(value) => {
-                52.write(output)?;
-                value.write(output)
-            }
-            Self::Delay(value) => {
-                53.write(output)?;
-                value.write(output)
-            }
-            Self::Loop(value) => {
-                54.write(output)?;
-                value.write(output)
-            }
-            Self::AutoPlay(value) => {
-                55.write(output)?;
-                value.write(output)
-            }
-            Self::LifetimeMin(value) => {
-                56.write(output)?;
-                value.write(output)
-            }
-            Self::LifetimeMax(value) => {
-                57.write(output)?;
-                value.write(output)
-            }
-            Self::SimulationSpace(value) => {
-                58.write(output)?;
-                value.write(output)
-            }
-            Self::Rate(value) => {
-                59.write(output)?;
-                value.write(output)
-            }
-            Self::Burst(value) => {
-                60.write(output)?;
-                value.write(output)
-            }
-            Self::EmitterShape(value) => {
-                61.write(output)?;
-                value.write(output)
-            }
-            Self::EmitterWidth(value) => {
-                62.write(output)?;
-                value.write(output)
-            }
-            Self::EmitterHeight(value) => {
-                63.write(output)?;
-                value.write(output)
-            }
-            Self::EmitterTotalAngle(value) => {
-                64.write(output)?;
-                value.write(output)
-            }
-            Self::SizeMin(value) => {
-                65.write(output)?;
-                value.write(output)
-            }
-            Self::SizeMax(value) => {
-                66.write(output)?;
-                value.write(output)
-            }
-            Self::SizeOverLifetime(value) => {
-                67.write(output)?;
-                value.write(output)
-            }
-            Self::StartSizeMultiplier(value) => {
-                68.write(output)?;
-                value.write(output)
-            }
-            Self::EndSizeMultiplier(value) => {
-                69.write(output)?;
-                value.write(output)
-            }
-            Self::SpeedMin(value) => {
-                71.write(output)?;
-                value.write(output)
-            }
-            Self::SpeedMax(value) => {
-                72.write(output)?;
-                value.write(output)
-            }
-            Self::SpeeLimit(value) => {
-                73.write(output)?;
-                value.write(output)
-            }
-            Self::SpeedDampen(value) => {
-                74.write(output)?;
-                value.write(output)
-            }
-            Self::RotationMin(value) => {
-                75.write(output)?;
-                value.write(output)
-            }
-            Self::RotationMax(value) => {
-                76.write(output)?;
-                value.write(output)
-            }
-            Self::Rotationspeed(value) => {
-                77.write(output)?;
-                value.write(output)
-            }
-            Self::ColourOverLifetime(value) => {
-                78.write(output)?;
-                value.write(output)
-            }
-            Self::StartColourMultiplier(value) => {
-                79.write(output)?;
-                value.write(output)
-            }
-            Self::EndColourMultiplier(value) => {
-                80.write(output)?;
-                value.write(output)
-            }
-            Self::GravityMultiplier(value) => {
-                81.write(output)?;
-                value.write(output)
-            }
-            Self::AnchorPos(value) => {
-                82.write(output)?;
-                value.write(output)
-            }
-            Self::MoonInnerRadius(value) => {
-                83.write(output)?;
-                value.write(output)
-            }
-            Self::MoonOffset(value) => {
-                84.write(output)?;
-                value.write(output)
-            }
-        }
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug)]
-pub struct Brush {
-    pub brush_id: i32,
-    pub spread: Vec2,
-    pub frequency: f32,
-    pub grid: BrushGrid,
-    pub objects: Vec<BrushObject>,
-}
-
-impl Read for Brush {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self {
-            brush_id: Read::read(input)?,
-            spread: Read::read(input)?,
-            frequency: Read::read(input)?,
-            grid: Read::read(input)?,
-            objects: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for Brush {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.brush_id.write(output)?;
-        self.spread.write(output)?;
-        self.frequency.write(output)?;
-        self.grid.write(output)?;
-        self.objects.write(output)
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Debug)]
-pub struct BrushObject {
-    pub entity_id: i32,
-    pub properties: Vec<ObjectProperty>,
-    pub weight: f32,
-    pub scale: f32,
-    pub rotation: f32,
-    pub flip_x: bool,
-    pub flip_y: bool,
-}
-
-impl Read for BrushObject {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self {
-            entity_id: Read::read(input)?,
-            properties: Read::read(input)?,
-            weight: Read::read(input)?,
-            scale: Read::read(input)?,
-            rotation: Read::read(input)?,
-            flip_x: Read::read(input)?,
-            flip_y: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for BrushObject {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.entity_id.write(output)?;
-        self.properties.write(output)?;
-        self.weight.write(output)?;
-        self.scale.write(output)?;
-        self.rotation.write(output)?;
-        self.flip_x.write(output)?;
-        self.flip_y.write(output)
-    }
-}
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub struct BrushGrid {
-    pub x: i32,
-    pub y: i32,
-}
-
-impl Read for BrushGrid {
-    fn read(input: &mut impl std::io::Read) -> Result<Self, Error> {
-        Ok(Self {
-            x: Read::read(input)?,
-            y: Read::read(input)?,
-        })
-    }
-}
-
-impl Write for BrushGrid {
-    fn write(&self, output: &mut impl std::io::Write) -> Result<(), Error> {
-        self.x.write(output)?;
-        self.y.write(output)
-    }
-}
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Clone, Debug)]
